# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SofaScore Scraper API",
    "description": "Versioned HTTP API of the SofaScore scraper. Single resources are returned as {\"data\": {...}}, collections as {\"data\": [...], \"page\": {\"limit\", \"next_cursor\"}}, errors as {\"error\": {\"code\", \"message\", \"details\", \"request_id\"}}. Messages are English; clients translate by code. When an access token is configured, every request needs Authorization: Bearer <token> or the session cookie.",
    "version": "1"
  },
  "paths": {
...
    "/api/v1/status": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Status of the service",
        "description": "Service state: versions, access to SofaScore, the request budget, the job that is running now, the live\nservice, what the data directory holds, the leases held right now and the optional features. When the\ndata directory cannot be read, `live` and `summary` are null and `storage_error` has the error code.",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/status/check": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Check the connection to SofaScore",
        "description": "Send one request to SofaScore now (the live list of football) and report whether it was answered. Only\ncalled when the user asks; it takes a turn of the shared request budget. A failed check answers 200 with\n`ok: false` and the `reason`: blocked, browser, rate_limited, network or upstream.",
        "operationId": "checkConnection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusCheckResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden_origin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sports": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List sports",
        "description": "The registered sports, in registry order, each with the detail slices that apply to it.",
        "operationId": "listSports",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportListResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sports/{slug}": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Get a sport",
        "operationId": "getSport",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SportResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sinks": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "List the output sinks",
        "description": "The sinks of the configuration (config file `[[sink]]` or SOFASCORE_SINKS), in their order, with the\nposition each has delivered up to and how far it lags behind the event log. Sinks are configured in the\nfile only; this route only reads. Credentials in webhook addresses are masked. Whether a process\ndelivers right now is `served` (the holder of the `sinks` lease is in `/status` `leases`).",
        "operationId": "listSinks",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SinkListResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "internal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "invalid_request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable error code; clients translate by this code."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "English text for logs and developers; never localised."
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details",
            "description": "Code-specific details."
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "Id of the request; also sent in the X-Request-Id header."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ApiError",
        "description": "The error object of every `/api/v1` error response."
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse"
      },
...
      "SinkListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SinkStatus"
            },
            "type": "array",
            "title": "Data"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "type": "object",
        "required": [
          "data",
          "page"
        ],
        "title": "SinkListResponse"
      },
...
      "SportListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Sport"
            },
            "type": "array",
            "title": "Data"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "type": "object",
        "required": [
          "data",
          "page"
        ],
        "title": "SportListResponse"
      },
      "SportResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Sport"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SportResponse"
      },
...
      "StatusCheckRequest": {
        "properties": {
          "target": {
            "type": "string",
            "const": "sofascore",
            "title": "Target",
            "description": "`sofascore`: one request for the live list of football."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target"
        ],
        "title": "StatusCheckRequest",
        "description": "What to check."
      },
      "StatusCheckResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/StatusCheck"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StatusCheckResponse"
      },
      "StatusResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "StatusResponse"
      },
...
    }
  }
}